requirements.txt
torch>=1.10
torchvision>=0.11
efficientnet-pytorch>=0.6.3
Pillow>=8.4.0
Flask>=2.0.0
Werkzeug>=2.0.0
